@@ -18,8 +18,9 @@
     "lodash": "^4.17.4",
     "moize": "^4.0.4",
     "namespace-constants": "^0.5.0",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
+    "prop-types": "^15.6.2",
+    "react": "^16.7.0",
+    "react-dom": "^16.7.0",
     "react-github-corner": "^2.1.0",
     "react-icons": "^2.2.7",
     "react-loader": "^2.4.2",
@@ -46,19 +47,20 @@
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-2": "^6.24.1",
     "babel-runtime": "^6.26.0",
-    "css-loader": "^0.28.9",
-    "eslint": "^4.16.0",
-    "eslint-config-rapid7": "^2.11.1",
-    "eslint-friendly-formatter": "^3.0.0",
-    "eslint-loader": "^1.9.0",
-    "file-loader": "^1.1.6",
-    "html-webpack-plugin": "^2.30.1",
+    "css-loader": "^2.1.0",
+    "eslint": "^5.13.0",
+    "eslint-config-rapid7": "^3.1.0",
+    "eslint-friendly-formatter": "^4.0.1",
+    "eslint-loader": "^2.1.2",
+    "file-loader": "^3.0.1",
+    "html-webpack-plugin": "^3.2.0",
     "lodash-webpack-plugin": "^0.11.4",
     "optimize-js-plugin": "^0.0.4",
-    "style-loader": "^0.20.1",
-    "url-loader": "^0.6.2",
-    "webpack": "^3.10.0",
-    "webpack-dev-server": "^2.11.1"
+    "style-loader": "^0.23.1",
+    "url-loader": "^1.1.2",
+    "webpack": "^4.29.3",
+    "webpack-cli": "^3.2.3",
+    "webpack-dev-server": "^3.1.14"
   },
   "homepage": "https://github.com/planttheidea/planttheidea.github.io#readme",
   "license": "UNLICENSED",
