@@ -1,5 +1,8 @@
 // external dependencies
-import {identity, identitySecond} from 'identitate';
+import {
+  identity,
+  identitySecond,
+} from 'identitate';
 import createConstants from 'namespace-constants';
 import {createAction} from 'redux-actions';
 
@@ -10,7 +13,7 @@ export const ACTION_TYPES = createConstants('repository', [
   'CLEAR_README',
   'GET_README_FAIL',
   'GET_README_PENDING',
-  'GET_README_SUCCESS'
+  'GET_README_SUCCESS',
 ]);
 
 export const clearReadme = createAction(ACTION_TYPES.CLEAR_README, () => {});
@@ -28,16 +31,14 @@ export const getReadmeSuccess = createAction(ACTION_TYPES.GET_README_SUCCESS, id
  * @param {string} repositoryName the name of the repository
  * @returns {function(function): Promise}
  */
-export const getReadme = (repositoryName) => {
-  return async (dispatch) => {
-    dispatch(getReadmePending(null, repositoryName));
+export const getReadme = (repositoryName) => async (dispatch) => {
+  dispatch(getReadmePending(null, repositoryName));
 
-    try {
-      const data = await reposApi.getRepositoryReadme(repositoryName);
+  try {
+    const data = await reposApi.getRepositoryReadme(repositoryName);
 
-      dispatch(getReadmeSuccess(data, repositoryName));
-    } catch (error) {
-      dispatch(getReadmeFail(error, repositoryName));
-    }
-  };
+    dispatch(getReadmeSuccess(data, repositoryName));
+  } catch (error) {
+    dispatch(getReadmeFail(error, repositoryName));
+  }
 };
