@@ -7,7 +7,7 @@ import {ACTION_TYPES} from 'actions/userProfileActions';
 export const INITIAL_STATE = {
   isLoadingUserProfile: false,
   userProfile: {},
-  userProfileError: null
+  userProfileError: null,
 };
 
 /**
@@ -20,13 +20,11 @@ export const INITIAL_STATE = {
  * @param {Error} payload the payload of the action
  * @returns {Object} the updated state of the reducer
  */
-const getUserProfileFail = (state, {payload: userProfileError}) => {
-  return {
-    ...state,
-    isLoadingUserProfile: false,
-    userProfileError
-  };
-};
+const getUserProfileFail = (state, {payload: userProfileError}) => ({
+  ...state,
+  isLoadingUserProfile: false,
+  userProfileError,
+});
 
 /**
  * @function getUserProfilePending
@@ -37,12 +35,10 @@ const getUserProfileFail = (state, {payload: userProfileError}) => {
  * @param {Object} state the current state of the reducer
  * @returns {Object} the updated state of the reducer
  */
-const getUserProfilePending = (state) => {
-  return {
-    ...state,
-    isLoadingUserProfile: true
-  };
-};
+const getUserProfilePending = (state) => ({
+  ...state,
+  isLoadingUserProfile: true,
+});
 
 /**
  * @function getUserProfileSuccess
@@ -54,20 +50,18 @@ const getUserProfilePending = (state) => {
  * @param {Object} payload the payload of the action
  * @returns {Object} the updated state of the reducer
  */
-const getUserProfileSuccess = (state, {payload: userProfile}) => {
-  return {
-    ...state,
-    isLoadingUserProfile: false,
-    userProfile,
-    userProfileError: null
-  };
-};
+const getUserProfileSuccess = (state, {payload: userProfile}) => ({
+  ...state,
+  isLoadingUserProfile: false,
+  userProfile,
+  userProfileError: null,
+});
 
 export default handleActions(
   {
     [ACTION_TYPES.GET_USER_PROFILE_FAIL]: getUserProfileFail,
     [ACTION_TYPES.GET_USER_PROFILE_PENDING]: getUserProfilePending,
-    [ACTION_TYPES.GET_USER_PROFILE_SUCCESS]: getUserProfileSuccess
+    [ACTION_TYPES.GET_USER_PROFILE_SUCCESS]: getUserProfileSuccess,
   },
   INITIAL_STATE
 );
