@@ -8,36 +8,15 @@ const defaultConfig = require('./webpack.config');
 module.exports = Object.assign({}, defaultConfig, {
   devtool: undefined,
 
+  mode: 'production',
+
   plugins: defaultConfig.plugins.concat([
     new webpack.LoaderOptionsPlugin({
       debug: false,
-      minimize: true
-    }),
-    new webpack.optimize.OccurrenceOrderPlugin(),
-    new webpack.optimize.ModuleConcatenationPlugin(),
-    new webpack.optimize.UglifyJsPlugin({
-      comments: false,
-      compress: {
-        booleans: true,
-        comparisons: true,
-        conditionals: true,
-        dead_code: true,
-        drop_console: true,
-        drop_debugger: true,
-        evaluate: true,
-        if_return: true,
-        join_vars: true,
-        loops: true,
-        properties: true,
-        screw_ie8: true,
-        sequences: true,
-        unused: true,
-        warnings: false
-      },
-      sourceMap: false
+      minimize: true,
     }),
     new OptimizeJsPlugin({
-      sourceMap: false
-    })
-  ])
+      sourceMap: false,
+    }),
+  ]),
 });
